30% from the wishlist and fix the userId

+++ server/controller/userController.js
@@ -1,3 +1,241 @@
+// const { User } = require("../indexdatabase");
+// const bcrypt = require("bcryptjs");
+// const jwt = require("jsonwebtoken");
+// require("dotenv").config();
+// const secret = process.env.JWT_SECRET;
+
+// // Get all users
+// const getAllusers = async (req, res) => {
+//   try {
+//     const users = await User.findAll();
+//     res.status(200).json(users);
+//   } catch (err) {
+//     res
+//       .status(500)
+//       .json({ message: "Failed to retrieve users", error: err.message });
+//   }
+// };
+
+// // Get one user by ID
+// const getOneUser = async (req, res) => {
+//   try {
+//     const oneUser = await User.findByPk(req.params.id);
+//     if (!oneUser) {
+//       return res.status(404).json({ message: "User not found" });
+//     }
+//     res.status(200).json(oneUser);
+//   } catch (err) {
+//     res
+//       .status(500)
+//       .json({ message: "Failed to retrieve user", error: err.message });
+//   }
+// };
+
+// // Delete a user
+// const deleteUser = async (req, res) => {
+//   try {
+//     const userId = req.params.id;
+//     console.log("User ID:", userId);
+//     const deleted = await User.destroy({ where: { userId: userId } });
+
+//     if (!deleted) {
+//       return res.status(404).json({ message: "User not found" });
+//     }
+
+//     res.status(200).json({ message: "Deleted user with ID: " + userId });
+//   } catch (error) {
+//     res.status(500).json({
+//       message: "An error occurred while deleting user",
+//       error: error.message,
+//     });
+//   }
+// };
+
+// // Sign up a new user
+// const singnIn = async (req, res) => {
+//   try {
+//     const {
+//       name,
+//       lastName,
+//       email,
+//       birthDate,
+//       adress,
+//       phoneNumber,
+//       password,
+//       role,
+//     } = req.body;
+
+//     const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+//     if (!emailRegex.test(email)) {
+//       return res.status(400).json({ message: "Invalid email format" });
+//     }
+
+//     // Check if the password meets the criteria
+//     const isPasswordValid = /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])[^_\s]{6,}$/.test(
+//       password
+//     );
+//     if (!isPasswordValid) {
+//       return res
+//         .status(400)
+//         .json({ message: "Password doesn't meet the criteria" });
+//     }
+
+//     // Check if the email is already used
+//     const existingUser = await User.findOne({ where: { email } });
+//     if (existingUser) {
+//       return res.status(400).json({ message: "Email is already used" });
+//     }
+
+//     // Create the new user
+//     const newUser = await User.create({
+//       name,
+//       lastName,
+//       email,
+//       birthDate,
+//       adress,
+//       phoneNumber,
+//       role,
+//       password: await bcrypt.hash(password, 10),
+//     });
+
+//     // Generate JWT this the parameter that we can include when we generate the token for the user
+//     const token = jwt.sign(
+//       { id: newUser.userId, email: newUser.email },
+//       secret,
+//       {
+//         expiresIn: "1h",
+//       }
+//     );
+
+//     return res.status(201).json({ token, message: "Sign In successful" });
+//   } catch (error) {
+//     console.error(error);
+//     return res
+//       .status(500)
+//       .json({ message: "Server Error", error: error.message });
+//   }
+// };
+
+// // Log in an existing user
+// const loginUser = async (req, res) => {
+//   try {
+//     const { email, password } = req.body;
+
+//     // Find user by email
+//     const user = await User.findOne({ where: { email } });
+//     if (!user) {
+//       return res.status(401).json({ message: "Invalid email or password" });
+//     }
+
+//     // Compare the password
+//     const isPasswordValid = await bcrypt.compare(password, user.password);
+//     if (!isPasswordValid) {
+//       return res.status(401).json({ message: "Invalid email or password" });
+//     }
+
+//     // Generate JWT
+//     const token = jwt.sign(
+//       {
+//         id: newUser.userId,
+//         name: newUser.name,
+//         lastName: newUser.lastName,
+//         phoneNumber: newUser.phoneNumber,
+//         email: newUser.email,
+//         adress: newUser.adress,
+//         birthDate: newUser.birthDate,
+//       },
+//       secret,
+//       { expiresIn: "1h" }
+//     );
+
+//     return res.json({ token });
+//   } catch (error) {
+//     console.error(error);
+//     return res
+//       .status(500)
+//       .json({ message: "Server Error", error: error.message });
+//   }
+// };
+
+// // Add a new user
+// const addUser = async (req, res) => {
+//   try {
+//     const {
+//       name,
+//       lastName,
+//       email,
+//       birthDate,
+//       adress,
+//       phoneNumber,
+//       password,
+//       role,
+//     } = req.body;
+
+//     const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+//     if (!emailRegex.test(email)) {
+//       return res.status(400).json({ message: "Invalid email format" });
+//     }
+
+//     const isPasswordValid = /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])[^_\s]{6,}$/.test(
+//       password
+//     );
+//     if (!isPasswordValid) {
+//       return res
+//         .status(400)
+//         .json({ message: "Password doesn't meet the criteria" });
+//     }
+
+//     const existingUser = await User.findOne({ where: { email } });
+//     if (existingUser) {
+//       return res.status(400).json({ message: "Email is already used" });
+//     }
+
+//     const hashedPassword = await bcrypt.hash(password, 10);
+//     const newUser = await User.create({
+//       name,
+//       lastName,
+//       email,
+//       birthDate,
+//       adress,
+//       phoneNumber,
+//       password: hashedPassword,
+//       role,
+//     });
+
+//     const token = jwt.sign(
+//       {
+//         id: newUser.userId,
+//         name: newUser.name,
+//         lastName: newUser.lastName,
+//         phoneNumber: newUser.phoneNumber,
+//         email: newUser.email,
+//         adress: newUser.adress,
+//         birthDate: newUser.birthDate,
+//       },
+//       secret,
+//       { expiresIn: "1h" }
+//     );
+
+//     return res
+//       .status(201)
+//       .json({ message: "User added successfully", newUser, token });
+//   } catch (error) {
+//     console.error(error);
+//     return res
+//       .status(500)
+//       .json({ message: "Server Error", error: error.message });
+//   }
+// };
+
+// module.exports = {
+//   getAllusers,
+//   getOneUser,
+//   deleteUser,
+//   singnIn,
+//   loginUser,
+//   addUser,
+// };
+
 const { User } = require("../indexdatabase");
 const bcrypt = require("bcryptjs");
 const jwt = require("jsonwebtoken");
@@ -8,9 +246,11 @@ const secret = process.env.JWT_SECRET;
 const getAllusers = async (req, res) => {
   try {
     const users = await User.findAll();
-    res.status(200).send(users);
+    res.status(200).json(users);
   } catch (err) {
-    res.status(404).send(err);
+    res
+      .status(500)
+      .json({ message: "Failed to retrieve users", error: err.message });
   }
 };
 
@@ -18,25 +258,34 @@ const getAllusers = async (req, res) => {
 const getOneUser = async (req, res) => {
   try {
     const oneUser = await User.findByPk(req.params.id);
-    res.status(200).send(oneUser);
+    if (!oneUser) {
+      return res.status(404).json({ message: "User not found" });
+    }
+    res.status(200).json(oneUser);
   } catch (err) {
-    res.status(404).send(err);
+    res
+      .status(500)
+      .json({ message: "Failed to retrieve user", error: err.message });
   }
 };
 
+// Delete a user
 const deleteUser = async (req, res) => {
   try {
-    let userId = req.params.id;
-    await User.destroy({
-      where: {
-        id: userId,
-      },
-    });
-    res.status(200).send("Deleted user with ID:" + userId);
+    const userId = req.params.id;
+    console.log("User ID:", userId);
+    const deleted = await User.destroy({ where: { userId: userId } });
+
+    if (!deleted) {
+      return res.status(404).json({ message: "User not found" });
+    }
+
+    res.status(200).json({ message: "Deleted user with ID: " + userId });
   } catch (error) {
-    res
-      .status(500)
-      .send({ error: "An error occurred while deleting user item" });
+    res.status(500).json({
+      message: "An error occurred while deleting user",
+      error: error.message,
+    });
   }
 };
 
@@ -55,31 +304,28 @@ const singnIn = async (req, res) => {
     } = req.body;
 
     const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-    const isEmailValid = emailRegex.test(email);
-    if (!isEmailValid) {
-      return res.status(400).send({ message: "Invalid email format" });
+    if (!emailRegex.test(email)) {
+      return res.status(400).json({ message: "Invalid email format" });
     }
 
     // Check if the password meets the criteria
     const isPasswordValid = /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])[^_\s]{6,}$/.test(
       password
     );
+    if (!isPasswordValid) {
+      return res
+        .status(400)
+        .json({ message: "Password doesn't meet the criteria" });
+    }
 
     // Check if the email is already used
     const existingUser = await User.findOne({ where: { email } });
     if (existingUser) {
-      return res.status(200).send({ message: "Email is already used" });
-    }
-
-    // Validate the password criteria
-    if (!isPasswordValid) {
-      return res
-        .status(400)
-        .send({ message: "Password doesn't meet the criteria" });
+      return res.status(400).json({ message: "Email is already used" });
     }
 
     // Create the new user
-    const newUser = await User.create({
+    const user = await User.create({
       name,
       lastName,
       email,
@@ -90,24 +336,29 @@ const singnIn = async (req, res) => {
       password: await bcrypt.hash(password, 10),
     });
 
-    // Generate JWT
+    // Generate JWT this the parameter that we can include when we generate the token for the user
     const token = jwt.sign(
       {
-        id: newUser.id,
-        name: newUser.name,
-        lastName: newUser.lastName,
-        email: newUser.email,
-        birthDate: newUser.birthDate,
-        adress: newUser.adress,
-        phoneNumber: newUser.phoneNumber,
-        role: newUser.role,
+        id: user.userId, // Use user.userId instead of newUser.userId
+        name: user.name,
+        lastName: user.lastName,
+        phoneNumber: user.phoneNumber,
+        email: user.email,
+        adress: user.adress,
+        birthDate: user.birthDate,
       },
-      secret
+      secret,
+      {
+        expiresIn: "1h",
+      }
     );
-    return res.status(201).send({ token, message: "Sign In successful" });
+
+    return res.status(201).json({ token, message: "Sign In successful" });
   } catch (error) {
-    console.log(error);
-    return res.status(500).send({ message: "Server Error", error });
+    console.error(error);
+    return res
+      .status(500)
+      .json({ message: "Server Error", error: error.message });
   }
 };
 
@@ -119,36 +370,41 @@ const loginUser = async (req, res) => {
     // Find user by email
     const user = await User.findOne({ where: { email } });
     if (!user) {
-      return res.status(404).send({ message: "Invalid email or password" });
+      return res.status(401).json({ message: "Invalid email or password" });
     }
 
     // Compare the password
     const isPasswordValid = await bcrypt.compare(password, user.password);
     if (!isPasswordValid) {
-      return res.status(404).send({ message: "Invalid email or password" });
+      return res.status(401).json({ message: "Invalid email or password" });
     }
 
     // Generate JWT
     const token = jwt.sign(
       {
-        id: user.id,
+        id: user.userId,
         name: user.name,
         lastName: user.lastName,
-        email: user.email,
-        birthDate: user.birthDate,
-        adress: user.adress,
         phoneNumber: user.phoneNumber,
+        email: user.email,
+        adress: user.adress,
+        birthDate: user.birthDate,
         role: user.role,
       },
-      secret
+      secret,
+      { expiresIn: "1h" }
     );
-    return res.send({ token });
+
+    return res.json({ token });
   } catch (error) {
-    console.log(error);
-    return res.status(500).send({ message: "Server Error", error });
+    console.error(error);
+    return res
+      .status(500)
+      .json({ message: "Server Error", error: error.message });
   }
 };
 
+// Add a new user
 const addUser = async (req, res) => {
   try {
     const {
@@ -163,9 +419,8 @@ const addUser = async (req, res) => {
     } = req.body;
 
     const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-    const isEmailValid = emailRegex.test(email);
-    if (!isEmailValid) {
-      return res.status(400).send({ message: "Invalid email format" });
+    if (!emailRegex.test(email)) {
+      return res.status(400).json({ message: "Invalid email format" });
     }
 
     const isPasswordValid = /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])[^_\s]{6,}$/.test(
@@ -174,16 +429,16 @@ const addUser = async (req, res) => {
     if (!isPasswordValid) {
       return res
         .status(400)
-        .send({ message: "Password doesn't meet the criteria" });
+        .json({ message: "Password doesn't meet the criteria" });
     }
 
     const existingUser = await User.findOne({ where: { email } });
     if (existingUser) {
-      return res.status(200).send({ message: "Email is already used" });
+      return res.status(400).json({ message: "Email is already used" });
     }
 
     const hashedPassword = await bcrypt.hash(password, 10);
-    const newUser = await User.create({
+    const user = await User.create({
       name,
       lastName,
       email,
@@ -194,12 +449,29 @@ const addUser = async (req, res) => {
       role,
     });
 
+    const token = jwt.sign(
+      {
+        id: user.userId,
+        name: user.name,
+        lastName: user.lastName,
+        phoneNumber: user.phoneNumber,
+        email: user.email,
+        adress: user.adress,
+        birthDate: user.birthDate,
+        role: user.role,
+      },
+      secret,
+      { expiresIn: "1h" }
+    );
+
     return res
       .status(201)
-      .send({ message: "User added successfully", newUser });
+      .json({ message: "User added successfully", user, token });
   } catch (error) {
-    console.log(error);
-    return res.status(500).send({ message: "Server Error", error });
+    console.error(error);
+    return res
+      .status(500)
+      .json({ message: "Server Error", error: error.message });
   }
 };
